requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
aiosqlite==0.19.0
requests==2.31.0 